Backend/requirements.txt
Django==4.2.18
djangorestframework==3.15.2
django-cors-headers==4.6.0
mysqlclient==2.2.6
